Create Error 404 job role not found - Impact of  Generative AI
Many of us are wondering if generative AI will be friend or foe to our roles. It is a topic for DevOps community who has been at forefront when it comes to onboarding new technologies. Our DevOps communities have been enabling and fostering the collaborative cultural shift, hand in hand evolving roles and responsibilities for software professionals. In this article we will try to oversee the indulgence of generative AI in the key roles when it comes to DevOps and try to outline the impact on key jobs, outlook and how do we prepare ourselves for the shift. 

Let us start from some basic roles, and their broad area of responsibility in the DevOps era. This section will enable us with the common vocabulary for the next section, where we delve into the impact with key trends in making.